move generator to dbc/
