+++ opendbc/dbc/generator/chrysler/_stellantis_common.dbc
