+++ opendbc/dbc/generator/chrysler/_stellantis_common_ram.py
