+++ opendbc/dbc/generator/chrysler/chrysler_pacifica_2017_hybrid.dbc
