+++ opendbc/dbc/generator/generator.py
