+++ opendbc/dbc/generator/gm/gm_global_a_powertrain.dbc
