+++ opendbc/dbc/generator/honda/_bosch_2018.dbc
