+++ opendbc/dbc/generator/honda/_community.dbc
