+++ opendbc/dbc/generator/honda/_honda_common.dbc
