+++ opendbc/dbc/generator/honda/_nidec_common.dbc
