+++ opendbc/dbc/generator/honda/_steering_sensors_a.dbc
