+++ opendbc/dbc/generator/honda/honda_civic_ex_2022_can.dbc
