+++ opendbc/dbc/generator/honda/honda_civic_hatchback_ex_2017_can.dbc
