+++ opendbc/dbc/generator/honda/honda_clarity_hybrid_2018_can.dbc
