+++ opendbc/dbc/generator/honda/honda_crv_touring_2016_can.dbc
