+++ opendbc/dbc/generator/honda/honda_insight_ex_2019_can.dbc
