+++ opendbc/dbc/generator/hyundai/hyundai_kia_mando_corner_radar.py
