+++ opendbc/dbc/generator/hyundai/hyundai_kia_mando_front_radar.py
