+++ opendbc/dbc/generator/nissan/_nissan_common.dbc
