+++ opendbc/dbc/generator/subaru/_subaru_global.dbc
