+++ opendbc/dbc/generator/subaru/_subaru_preglobal_2015.dbc
