+++ opendbc/dbc/generator/subaru/subaru_global_2020_hybrid.dbc
