+++ opendbc/dbc/generator/tesla/.gitignore
