+++ opendbc/dbc/generator/tesla/radar_common.py
