+++ opendbc/dbc/generator/toyota/toyota_nodsu_pt.dbc
